user-weight as floating-point
Code supported it halfway but then ignored the fractional part. Now full floating-point is supported in user wieght, eg "[3.14159265w]someword"

+++ Query.h
@@ -237,7 +237,7 @@ class QueryWord {
 	class QueryTerm *m_queryWordTerm;
 
 	// user defined weights
-	int32_t m_userWeightForWord;
+	float m_userWeightForWord;
 	float m_userWeightForPhrase;
 
 	bool m_queryOp;
